+++ src/output/mod.rs
@@ -5,6 +5,7 @@ pub(crate) mod display;
 pub mod lambdify;
 pub(crate) mod latex;
 pub mod lean;
+pub mod lean_proof;
 pub mod parse;
 pub(crate) mod pretty;
 pub mod tree;